+++ test/serving_scaler_test.dart
@@ -14,7 +14,7 @@ void main() {
     List<Ingredient> newIngredients = scaler.amountScaler(servings);
 
     for (int i = 0; i < ingredients.length; i++) {
-      int amount = ingredients[i].amount;
+      double amount = ingredients[i].amount;
       amount = amount * servings;
       ingredients[i].setAmount(amount);
     }
